LPS-115742 Move logic to ContentDashboardDropdownItemsProvider
