+++ modules/apps/content-dashboard/content-dashboard-web/src/main/java/com/liferay/content/dashboard/web/internal/display/context/ContentDashboardAdminDisplayContext.java
@@ -15,35 +15,28 @@
 package com.liferay.content.dashboard.web.internal.display.context;
 
 import com.liferay.content.dashboard.web.internal.item.ContentDashboardItem;
+import com.liferay.content.dashboard.web.internal.servlet.taglib.util.ContentDashboardDropdownItemsProvider;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem;
-import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemList;
 import com.liferay.item.selector.ItemSelector;
 import com.liferay.item.selector.criteria.UUIDItemSelectorReturnType;
 import com.liferay.portal.kernel.dao.search.SearchContainer;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
-import com.liferay.portal.kernel.portlet.PortletURLUtil;
 import com.liferay.portal.kernel.portlet.RequestBackedPortletURLFactory;
 import com.liferay.portal.kernel.portlet.RequestBackedPortletURLFactoryUtil;
 import com.liferay.portal.kernel.util.ArrayUtil;
-import com.liferay.portal.kernel.util.Http;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.users.admin.item.selector.UserItemSelectorCriterion;
 
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.Locale;
 
 import javax.portlet.PortletURL;
-
-import javax.servlet.http.HttpServletRequest;
 
 /**
  * @author Cristina González
@@ -51,22 +44,19 @@ import javax.servlet.http.HttpServletRequest;
 public class ContentDashboardAdminDisplayContext {
 
 	public ContentDashboardAdminDisplayContext(
-		Http http, ItemSelector itemSelector, Language language,
-		LiferayPortletRequest liferayPortletRequest,
+		ContentDashboardDropdownItemsProvider
+			contentDashboardDropdownItemsProvider,
+		ItemSelector itemSelector, LiferayPortletRequest liferayPortletRequest,
 		LiferayPortletResponse liferayPortletResponse, Portal portal,
 		SearchContainer<ContentDashboardItem<?>> searchContainer) {
 
-		_http = http;
+		_contentDashboardDropdownItemsProvider =
+			contentDashboardDropdownItemsProvider;
 		_itemSelector = itemSelector;
-		_language = language;
 		_liferayPortletRequest = liferayPortletRequest;
 		_liferayPortletResponse = liferayPortletResponse;
 		_portal = portal;
 		_searchContainer = searchContainer;
-
-		_currentURL = String.valueOf(
-			PortletURLUtil.getCurrent(
-				_liferayPortletRequest, _liferayPortletResponse));
 	}
 
 	public List<Long> getAuthorIds() {
@@ -108,47 +98,8 @@ public class ContentDashboardAdminDisplayContext {
 	public List<DropdownItem> getDropdownItems(
 		ContentDashboardItem contentDashboardItem) {
 
-		Locale locale = _portal.getLocale(_liferayPortletRequest);
-
-		return DropdownItemList.of(
-			() -> {
-				HttpServletRequest httpServletRequest =
-					_portal.getHttpServletRequest(_liferayPortletRequest);
-
-				if (!contentDashboardItem.isViewURLEnabled(
-						httpServletRequest)) {
-
-					return null;
-				}
-
-				DropdownItem dropdownItem = new DropdownItem();
-
-				dropdownItem.setHref(
-					_getURLWithBackURL(
-						contentDashboardItem.getViewURL(httpServletRequest)));
-				dropdownItem.setLabel(_language.get(locale, "view"));
-
-				return dropdownItem;
-			},
-			() -> {
-				HttpServletRequest httpServletRequest =
-					_portal.getHttpServletRequest(_liferayPortletRequest);
-
-				if (!contentDashboardItem.isEditURLEnabled(
-						httpServletRequest)) {
-
-					return null;
-				}
-
-				DropdownItem dropdownItem = new DropdownItem();
-
-				dropdownItem.setHref(
-					_getURLWithBackURL(
-						contentDashboardItem.getEditURL(httpServletRequest)));
-				dropdownItem.setLabel(_language.get(locale, "edit"));
-
-				return dropdownItem;
-			});
+		return _contentDashboardDropdownItemsProvider.getDropdownItems(
+			contentDashboardItem);
 	}
 
 	public SearchContainer<ContentDashboardItem<?>> getSearchContainer() {
@@ -176,21 +127,10 @@ public class ContentDashboardAdminDisplayContext {
 		return _userId;
 	}
 
-	private String _getURLWithBackURL(String url) {
-		String backURL = ParamUtil.getString(_liferayPortletRequest, "backURL");
-
-		if (Validator.isNotNull(backURL)) {
-			return _http.setParameter(url, "p_l_back_url", backURL);
-		}
-
-		return _http.setParameter(url, "p_l_back_url", _currentURL);
-	}
-
 	private List<Long> _authorIds;
-	private final String _currentURL;
-	private final Http _http;
+	private final ContentDashboardDropdownItemsProvider
+		_contentDashboardDropdownItemsProvider;
 	private final ItemSelector _itemSelector;
-	private final Language _language;
 	private final LiferayPortletRequest _liferayPortletRequest;
 	private final LiferayPortletResponse _liferayPortletResponse;
 	private final Portal _portal;
